#63 add fastapi router, restructure project files
